refactor: Se separa la lógica de la galería en GalleryView y se reutiliza CarDetailView
